requirements.txt
Django==1.4.2
MySQL-python==1.2.4
psycopg2==2.5
South==0.7.6
dj-database-url==0.2.1
python-dateutil==2.1
django-extensions==1.1.1

